refactor: implement task ownership validation in attachment service
