+++ services/api/internal/service/attachment/service.go
@@ -3,6 +3,7 @@ package attachmentsvc
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"mime"
 	"path"
@@ -12,6 +13,7 @@ import (
 
 	"github.com/google/uuid"
 	attachmentdom "github.com/paca/api/internal/domain/attachment"
+	taskdom "github.com/paca/api/internal/domain/task"
 	"github.com/paca/api/internal/platform/storage"
 )
 
@@ -266,4 +268,30 @@ func sanitizeFileName(name string) string {
 		name = "file"
 	}
 	return name
+}
+
+// --- Task ownership checker --------------------------------------------------
+
+type taskOwnerChecker struct {
+	repo taskdom.TaskRepository
+}
+
+// NewTaskOwnerChecker returns a attachmentdom.TaskOwnerChecker that validates
+// task→project ownership via the task repository.
+func NewTaskOwnerChecker(repo taskdom.TaskRepository) attachmentdom.TaskOwnerChecker {
+	return &taskOwnerChecker{repo: repo}
+}
+
+func (c *taskOwnerChecker) TaskBelongsToProject(ctx context.Context, projectID, taskID uuid.UUID) error {
+	t, err := c.repo.FindTaskByID(ctx, taskID)
+	if err != nil {
+		if errors.Is(err, taskdom.ErrTaskNotFound) {
+			return attachmentdom.ErrTaskNotInProject
+		}
+		return err
+	}
+	if t.ProjectID != projectID {
+		return attachmentdom.ErrTaskNotInProject
+	}
+	return nil
 }
